Handle case with zero bytes remaining in stream
There was a missing case in the scene stream parsing where it did not account for cases where there could be zero bytes remaining in the block. This update handles that case by immediately returning the subclass after reading the base sequence details with no extra data. This prevents any issues of trying to read data from the stream when there isn't any left.